+++ NEWS.md
@@ -1,7 +1,9 @@
-# repana (development version)
-
 # NEWS repana
 
+## repana 2.2.0
+Add a `targets_structure` function to use the targets system instead
+ of the default used by repana.
+ 
 ## repana 2.1.0
 
 * Complete vignette to render reports